支付宝通道下单接口 支付宝通道支付回调接口

+++ src/main/java/com/example/jiuzhou/user/model/AbpParkadeAccessDetail.java
@@ -50,7 +50,7 @@ public class AbpParkadeAccessDetail {
     private Date CarOutTime;
 
     @Column(name="StopTime")
-    private Integer StopTime;
+    private long StopTime;
 
     @Column(name="guid")
     private String guid;
